[#151]Feat(badge): BadgeService 생성하고 배지 처리하는 기능을 이벤트리스너에서 BadgeService로 이동"
